# SerializedCellVec improve serialization performance