SerializedCellVec improve serialization performance