@kitaisreal 2023-07-03
Run performance test Scheme::TSerializedCellVec in `scheme_tablecell_ut.cpp`.
Iterations 10000000 cells size 12:
```
Before:

Serialize: 0.580962s
Cells constructor: 1.067759s
Parse: 0.444905s
Copy: 0.183273s
Move: 0.022736s

After:

Serialize: 0.563385s
Cells constructor: 0.862292s
Parse: 0.379019s
Copy: 0.180011s
Move: 0.022593s
```

Iterations 10000000 cells size 60:
```
Before:

Iterations 10000000 cells size 60
Serialize: 1.936293s
Cells constructor: 3.550640s
Parse: 2.201326s
Copy: 0.223238s
Move: 0.023384s

After:

Iterations 10000000 cells size 60
Serialize: 1.838559s
Cells constructor: 3.261016s
Parse: 1.419611s
Copy: 0.203151s
Move: 0.022990s
```